+++ lib/karchive/testdata/small.zip
