--- lib/kzip/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["unzip.go"],
-    importpath = "github.com/enfabrica/enkit/lib/kzip",
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["unzip_test.go"],
-    data = glob(["testdata/**"]),
-    embed = [":go_default_library"],
-    deps = [
-        "//lib/errdiff:go_default_library",
-        "//lib/testutil:go_default_library",
-        "@com_github_prashantv_gostub//:go_default_library",
-        "@com_github_stretchr_testify//assert:go_default_library",
-    ],
-)
